added some remaining functions

+++ ft_strtrim.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strtrim.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ydred <ydred@student.42.fr>                +#+  +:+       +#+        */
+/*   By: nmillier <nmillier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 15:54:05 by nmillier          #+#    #+#             */
-/*   Updated: 2023/11/02 17:10:56 by ydred            ###   ########.fr       */
+/*   Updated: 2023/11/03 13:16:02 by nmillier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ char	*ft_strtrim(char const *s1, char const *set)
 		start++;
 	while (end >= start && ft_charset(s1[end], (char *) set))
 		end--;
-	newstring = ft_substr(s1, start, end - start + 1);
+	newstring = ft_substr(s1, (unsigned int) start, (size_t) (end - start + 1));
 	if (newstring == NULL)
 		return (NULL);
 	return (newstring);
